+++ src/main/java/seedu/address/logic/parser/room/RoomCliSyntax.java
@@ -8,6 +8,6 @@ import seedu.address.logic.parser.Prefix;
 public class RoomCliSyntax {
 
     /* Prefix definitions */
-    public static final Prefix PREFIX_ROOM_NUMBER = new Prefix("rn/");
-    public static final Prefix PREFIX_PATIENT_NAME = new Prefix("pn/");
+    public static final Prefix PREFIX_ROOM_NUMBER = new Prefix("r/");
+    public static final Prefix PREFIX_PATIENT_NAME = new Prefix("p/");
 }
